Lots of changes.
Explicitly insert CSS into the webpage by tab Id. Clicking on the dev console and changing a setting on the popup (or options page) would attempt to insert the CSS into the console window and silently yet dysfunctionally error because it happens in the background script.

Added a resizable record button in the options page.