Features and updates
Fix typo (extra }) in initMontage().
Made numbers input larger on Option page.
Changed default grid maximum size (and flash alarm) to 20 from 10.
Made it so clearing the storage for the extension will also reload everything, even the ZoneMinder page.
Added the functionality to go directly back to the console from the montage page by clicking the right and left mouse buttons at the same time. And if the montage page was in F11 fullscreen mode, make it back to normal. Will adjust this  to go back to whatever state it was in.